chore: split state modules
