+++ docs/deep-dives/extractors-layer.md
@@ -19,8 +19,8 @@ Prerequisites:
 2. [Stream Layer (`binary/stream.py`)](stream-layer.md)
 3. [Parser Layer (`parser.py`)](parser-layer.md)
 4. [SendTable Layer (`schema/sendtable.py`)](sendtable-layer.md)
-5. [State Reconstruction Layer (`string_table.py` + `entities.py`)](state-layer.md)
-6. [Event Normalization Layer (`game_events.py` + `combatlog.py`)](event-layer.md)
+5. [State Reconstruction Layer (`state/string_table.py` + `state/entities.py`)](state-layer.md)
+6. [Event Normalization Layer (`state/game_events.py` + `combatlog.py`)](event-layer.md)
 
 ## Why this is the next layer
