+++ docs/reference/game_events.md
@@ -12,7 +12,7 @@ See also: [How Proto Parsing Works](../cookbook/proto-parsing-pipeline.md)
 
 ## Generated API
 
-## `gem.game_events.GameEventManager`
+## `gem.state.game_events.GameEventManager`
 
 ### `GameEventManager`
 
@@ -22,7 +22,7 @@ class GameEventManager
 
 Manages game event schema registration and handler dispatch.
 
-Source: [src/gem/game_events.py:146](https://github.com/whanyu1212/gem-dota/blob/main/src/gem/game_events.py#L146)
+Source: [src/gem/state/game_events.py:146](https://github.com/whanyu1212/gem-dota/blob/main/src/gem/state/game_events.py#L146)
 
 #### Methods
 
@@ -32,7 +32,7 @@ Signature: `def GameEventManager.register_schema(self, schema_dict: dict[str, An
 
 Register an event schema from a dict (e.g. from CSVCMsg_GameEventList).
 
-Source: [src/gem/game_events.py:160](https://github.com/whanyu1212/gem-dota/blob/main/src/gem/game_events.py#L160)
+Source: [src/gem/state/game_events.py:160](https://github.com/whanyu1212/gem-dota/blob/main/src/gem/state/game_events.py#L160)
 
 ##### `has_event`
 
@@ -40,7 +40,7 @@ Signature: `def GameEventManager.has_event(self, name: str) -> bool`
 
 Return True if an event schema with the given name is registered.
 
-Source: [src/gem/game_events.py:178](https://github.com/whanyu1212/gem-dota/blob/main/src/gem/game_events.py#L178)
+Source: [src/gem/state/game_events.py:178](https://github.com/whanyu1212/gem-dota/blob/main/src/gem/state/game_events.py#L178)
 
 ##### `on_game_event`
 
@@ -48,7 +48,7 @@ Signature: `def GameEventManager.on_game_event(self, name: str, handler: GameEve
 
 Register a handler for the named event.
 
-Source: [src/gem/game_events.py:186](https://github.com/whanyu1212/gem-dota/blob/main/src/gem/game_events.py#L186)
+Source: [src/gem/state/game_events.py:186](https://github.com/whanyu1212/gem-dota/blob/main/src/gem/state/game_events.py#L186)
 
 ##### `dispatch`
 
@@ -56,9 +56,9 @@ Signature: `def GameEventManager.dispatch(self, raw_event: Any) -> None`
 
 Dispatch a raw CMsgSource1LegacyGameEvent message to registered handlers.
 
-Source: [src/gem/game_events.py:195](https://github.com/whanyu1212/gem-dota/blob/main/src/gem/game_events.py#L195)
+Source: [src/gem/state/game_events.py:195](https://github.com/whanyu1212/gem-dota/blob/main/src/gem/state/game_events.py#L195)
 
-## `gem.game_events.GameEvent`
+## `gem.state.game_events.GameEvent`
 
 ### `GameEvent`
 
@@ -68,7 +68,7 @@ class GameEvent
 
 A decoded game event instance.
 
-Source: [src/gem/game_events.py:40](https://github.com/whanyu1212/gem-dota/blob/main/src/gem/game_events.py#L40)
+Source: [src/gem/state/game_events.py:40](https://github.com/whanyu1212/gem-dota/blob/main/src/gem/state/game_events.py#L40)
 
 #### Methods
 
@@ -78,7 +78,7 @@ Signature: `def GameEvent.get_string(self, name: str) -> tuple[str, str | None]`
 
 Return (value, None) as str, or ('', error) on failure.
 
-Source: [src/gem/game_events.py:65](https://github.com/whanyu1212/gem-dota/blob/main/src/gem/game_events.py#L65)
+Source: [src/gem/state/game_events.py:65](https://github.com/whanyu1212/gem-dota/blob/main/src/gem/state/game_events.py#L65)
 
 ##### `get_float`
 
@@ -86,7 +86,7 @@ Signature: `def GameEvent.get_float(self, name: str) -> tuple[float, str | None]
 
 Return (value, None) as float, or (0.0, error) on failure.
 
-Source: [src/gem/game_events.py:80](https://github.com/whanyu1212/gem-dota/blob/main/src/gem/game_events.py#L80)
+Source: [src/gem/state/game_events.py:80](https://github.com/whanyu1212/gem-dota/blob/main/src/gem/state/game_events.py#L80)
 
 ##### `get_int32`
 
@@ -94,7 +94,7 @@ Signature: `def GameEvent.get_int32(self, name: str) -> tuple[int, str | None]`
 
 Return (value, None) as int32 (long/short/byte), or (0, error).
 
-Source: [src/gem/game_events.py:94](https://github.com/whanyu1212/gem-dota/blob/main/src/gem/game_events.py#L94)
+Source: [src/gem/state/game_events.py:94](https://github.com/whanyu1212/gem-dota/blob/main/src/gem/state/game_events.py#L94)
 
 ##### `get_bool`
 
@@ -102,7 +102,7 @@ Signature: `def GameEvent.get_bool(self, name: str) -> tuple[bool, str | None]`
 
 Return (value, None) as bool, or (False, error) on failure.
 
-Source: [src/gem/game_events.py:114](https://github.com/whanyu1212/gem-dota/blob/main/src/gem/game_events.py#L114)
+Source: [src/gem/state/game_events.py:114](https://github.com/whanyu1212/gem-dota/blob/main/src/gem/state/game_events.py#L114)
 
 ##### `get_uint64`
 
@@ -110,9 +110,9 @@ Signature: `def GameEvent.get_uint64(self, name: str) -> tuple[int, str | None]`
 
 Return (value, None) as uint64, or (0, error) on failure.
 
-Source: [src/gem/game_events.py:128](https://github.com/whanyu1212/gem-dota/blob/main/src/gem/game_events.py#L128)
+Source: [src/gem/state/game_events.py:128](https://github.com/whanyu1212/gem-dota/blob/main/src/gem/state/game_events.py#L128)
 
-## `gem.game_events.GameEventSchema`
+## `gem.state.game_events.GameEventSchema`
 
 ### `GameEventSchema`
 
@@ -122,7 +122,7 @@ class GameEventSchema
 
 Schema for a single game event type.
 
-Source: [src/gem/game_events.py:26](https://github.com/whanyu1212/gem-dota/blob/main/src/gem/game_events.py#L26)
+Source: [src/gem/state/game_events.py:26](https://github.com/whanyu1212/gem-dota/blob/main/src/gem/state/game_events.py#L26)
 
 #### Dataclass fields
