+++ scripts/generate_vitepress_api_reference.py
@@ -33,7 +33,11 @@ TARGETS_BY_PAGE: dict[str, list[str]] = {
     "combatlog.md": ["gem.combatlog.CombatLogProcessor", "gem.combatlog.CombatLogEntry"],
     "constants.md": ["gem.constants"],
     "dataframes.md": ["gem.dataframes"],
-    "entities.md": ["gem.entities.EntityOp", "gem.entities.Entity", "gem.entities.EntityManager"],
+    "entities.md": [
+        "gem.state.entities.EntityOp",
+        "gem.state.entities.Entity",
+        "gem.state.entities.EntityManager",
+    ],
     "field_decoder.md": [
         "gem.schema.field_decoder.find_decoder",
         "gem.schema.field_decoder.find_decoder_by_base_type",
@@ -45,9 +49,9 @@ TARGETS_BY_PAGE: dict[str, list[str]] = {
         "gem.schema.field_path.FieldPathOp",
     ],
     "game_events.md": [
-        "gem.game_events.GameEventManager",
-        "gem.game_events.GameEvent",
-        "gem.game_events.GameEventSchema",
+        "gem.state.game_events.GameEventManager",
+        "gem.state.game_events.GameEvent",
+        "gem.state.game_events.GameEventSchema",
     ],
     "match_builder.md": ["gem.match_builder"],
     "models.md": ["gem.models"],
@@ -61,12 +65,12 @@ TARGETS_BY_PAGE: dict[str, list[str]] = {
     ],
     "stream.md": ["gem.binary.stream.DemoStream", "gem.binary.stream.OuterMessage"],
     "string_table.md": [
-        "gem.string_table.StringTables",
-        "gem.string_table.StringTable",
-        "gem.string_table.StringTableItem",
-        "gem.string_table.parse_string_table",
-        "gem.string_table.handle_create",
-        "gem.string_table.handle_update",
+        "gem.state.string_table.StringTables",
+        "gem.state.string_table.StringTable",
+        "gem.state.string_table.StringTableItem",
+        "gem.state.string_table.parse_string_table",
+        "gem.state.string_table.handle_create",
+        "gem.state.string_table.handle_update",
     ],
     "extractors/courier.md": ["gem.extractors.courier"],
     "extractors/draft.md": ["gem.extractors.draft"],
