+++ src/gem/state/__init__.py
@@ -1,0 +1,38 @@
+"""Replay state reconstruction primitives."""
+
+from gem.state.entities import (
+    ClassInfo,
+    Entity,
+    EntityHandler,
+    EntityManager,
+    EntityOp,
+    EntityTracker,
+)
+from gem.state.game_events import GameEvent, GameEventHandler, GameEventManager, GameEventSchema
+from gem.state.string_table import (
+    StringTable,
+    StringTableItem,
+    StringTables,
+    handle_create,
+    handle_update,
+    parse_string_table,
+)
+
+__all__ = [
+    "ClassInfo",
+    "Entity",
+    "EntityHandler",
+    "EntityManager",
+    "EntityOp",
+    "EntityTracker",
+    "GameEvent",
+    "GameEventHandler",
+    "GameEventManager",
+    "GameEventSchema",
+    "StringTable",
+    "StringTableItem",
+    "StringTables",
+    "handle_create",
+    "handle_update",
+    "parse_string_table",
+]
